+++ src/Client.h
@@ -113,6 +113,7 @@ public slots:
     // Authentication
     void loginWithPassword(QString deviceName, QString userId, QString password, QString serverAddress);
     void loginWithCiba(QString username,QString server,QString accessToken = "");
+    void cancelCibaLogin();
     bool hasValidUser();
     UserInformation userInformation();
     void userInformation(const QString &mxid);