+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+
+## [0.1.20]
+### Changed
+
+* CIBA Login improvment.
+* Cancel CIBA Login request.
+
+
 ## [0.1.19]
 ### Changed